Adding comments/javadoc to Cacheca files

+++ plugins/com.python.pydev.codecompletion/src/com/python/pydev/codecompletion/simpleassist/Cache.java
@@ -1,3 +1,12 @@
+/**
+ * Copyright (c) 2005-2013 by Appcelerator, Inc. All Rights Reserved.
+ * Licensed under the terms of the Eclipse Public License (EPL).
+ * Please see the license.txt included with this distribution for details.
+ * Any modifications to this file must keep this entire header intact.
+ * 
+ * Code for Cacheca model implementation for code completion included - Siddhika Cowlagi,  Vincent Hellendoorn , Premkumar T Devanbu
+ */
+
 package com.python.pydev.codecompletion.simpleassist;
 
 import java.io.BufferedReader;
@@ -99,7 +108,7 @@ public class Cache {
             Map<String, Integer> tokenCounts = getTokenCounts(prefix);
 
             // update the information of candidates from ngram model
-            for (int i = 0; i < (int) candidates.size(); ++i)
+            for (int i = 0; i < candidates.size(); ++i)
             {
                 // discount the probability first
                 candidates.get(i).mProb *= ngram_discount;
